[Testing] move test_dir to wholesale container
