+++ source/databricks/package/datamigration/migration_scripts/202211091500_Dummy_Migration.py
@@ -15,5 +15,5 @@
 from pyspark.sql import SparkSession
 
 
-def apply(spark: SparkSession, storage_account_name: str, storage_account_key: str):
+def apply(storage_account_name: str, storage_account_key: str, spark: SparkSession):
     print("Applying dummy migration")
